REP-188: Move postgres metrics classes into DatabaseMetricsService
